third_party/blink/web_tests/platform/mac/external/wpt/webrtc-encoded-transform/script-transform-generateKeyFrame-simulcast.https-expected.txt

This is a testharness.js-based test.
[FAIL] generateKeyFrame works with simulcast rids
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame for rid that was negotiated away fails
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
[FAIL] generateKeyFrame with rid after simulcast->unicast negotiation fails
  assert_equals: expected (string) "got frame" but got (object) object "[object Object]"
Harness: the test ran to completion.
